+++ fueoni_ver2/lib/screens/home_screen/home_screen.dart
@@ -2,52 +2,43 @@ import 'dart:async';
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:fueoni_ver2/hooks/use_location.dart';
 import 'package:fueoni_ver2/screens/home_screen/tabs/profile_settings/account_settings_screen.dart';
 import 'package:fueoni_ver2/screens/home_screen/tabs/room_settings/room_settings.dart';
 
-class HomeScreen extends StatefulWidget {
+class HomeScreen extends HookWidget {
   const HomeScreen({Key? key}) : super(key: key);
 
   @override
-  State<HomeScreen> createState() => _HomeScreenState();
-}
-
-class _HomeScreenState extends State<HomeScreen> {
-  int _selectedIndex = 0;
-  StreamSubscription<User?>? _authSubscription;
-
-  @override
   Widget build(BuildContext context) {
+    useLocationPermissionCheck(context);
+    final selectedIndex = useState(0);
+    final authSubscription = useState<StreamSubscription<User?>?>(null);
+
+    useEffect(() {
+      authSubscription.value =
+          FirebaseAuth.instance.authStateChanges().listen((User? user) {
+        if (user == null) {
+          Navigator.of(context).pushNamedAndRemoveUntil('/', (route) => false);
+        }
+      });
+      return () {
+        authSubscription.value?.cancel();
+      };
+    }, const []);
+
     return Scaffold(
-      body: _buildSelectedScreen(),
-      bottomNavigationBar: _buildBottomNavigationBar(),
+      body: _buildSelectedScreen(selectedIndex.value),
+      bottomNavigationBar: _buildBottomNavigationBar(selectedIndex),
     );
   }
 
-  @override
-  void dispose() {
-    _authSubscription?.cancel();
-    super.dispose();
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    _authSubscription =
-        FirebaseAuth.instance.authStateChanges().listen((User? user) {
-      if (user == null) {
-        Navigator.of(context).pushNamedAndRemoveUntil('/', (route) => false);
-      }
-    });
-  }
-
-  NavigationBar _buildBottomNavigationBar() {
+  NavigationBar _buildBottomNavigationBar(ValueNotifier<int> selectedIndex) {
     return NavigationBar(
-      selectedIndex: _selectedIndex,
+      selectedIndex: selectedIndex.value,
       onDestinationSelected: (int index) {
-        setState(() {
-          _selectedIndex = index;
-        });
+        selectedIndex.value = index;
       },
       destinations: const [
         NavigationDestination(
@@ -62,9 +53,9 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 
-  Widget _buildSelectedScreen() {
+  Widget _buildSelectedScreen(int selectedIndex) {
     // 現在の選択に基づいて表示する画面を返します
-    switch (_selectedIndex) {
+    switch (selectedIndex) {
       case 0:
         return const RoomSettingsScreen();
       case 1:
