+++ fueoni_ver2/lib/screens/home_screen/tabs/profile_settings/account_settings_screen.dart
@@ -1,6 +1,7 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
+import 'package:fueoni_ver2/hooks/use_location.dart';
 import 'package:fueoni_ver2/main.dart';
 import 'package:fueoni_ver2/services/database/user.dart';
 import 'package:provider/provider.dart';
@@ -13,7 +14,7 @@ class AccountSettingsScreen extends HookWidget {
 
   @override
   Widget build(BuildContext context) {
-    // useLocationPermissionCheck(context);
+    useLocationPermissionCheck(context);
     final isExpanded = useState(false);
     final isLoading = useState(false);
